feat: create an atom to handle the qr-code-give-badge-address
