+++ src/components/01-atoms/index.ts
@@ -29,6 +29,7 @@ export * from "./CopyToClipboardButton";
 export * from "./CreatedByBlockful";
 export * from "./DropdownProfile";
 export * from "./InfoText";
+export * from "./QRCodeGiveBadge";
 export * from "./QRCodeScanner";
 export * from "./SignMessage";
 export * from "./TheFooterNavbar";
