+++ src/components/04-templates/CheckInSection.tsx
@@ -7,12 +7,12 @@ import {
   Flex,
   Divider,
 } from "@chakra-ui/react";
-import { useQRCode } from "next-qrcode";
 import { useAccount } from "wagmi";
 
 import {
   CopyToClipboardButton,
   CircleQuestion,
+  QRCodeGiveBadge,
   TheFooterNavbar,
   TheHeader,
   TicketIcon,
@@ -20,7 +20,6 @@ import {
 import { getEllipsedAddress } from "@/utils/formatters";
 
 export const CheckinSection = () => {
-  const { Canvas } = useQRCode();
   const { address, chain } = useAccount();
 
   return (
@@ -70,17 +69,8 @@ export const CheckinSection = () => {
             >
               {address && chain ? (
                 <>
-                  <Canvas
-                    text={address}
-                    options={{
-                      errorCorrectionLevel: "M",
-                      width: 250,
-                      color: {
-                        dark: "#F5FFFFDC",
-                        light: "#212223",
-                      },
-                    }}
-                  />
+                  <QRCodeGiveBadge />
+
                   <Flex
                     color="white"
                     className="justify-center items-center gap-2"
