+++ src/components/04-templates/ShareSection.tsx
@@ -7,32 +7,20 @@ import {
   Flex,
   Link,
 } from "@chakra-ui/react";
-import { useQRCode } from "next-qrcode";
 import { useAccount } from "wagmi";
 
 import {
   CopyToClipboardButton,
   BlockfulLogo,
+  QRCodeGiveBadge,
   TheFooterNavbar,
   TheHeader,
   TicketIcon,
 } from "@/components/01-atoms";
-import { useUrl } from "@/hooks/useURL";
 import { getEllipsedAddress } from "@/utils/formatters";
 
 export const ShareSection = () => {
-  const { Canvas } = useQRCode();
   const { address, chain } = useAccount();
-  const url = useUrl();
-  let prefixToGiveBadge: string = "http://localhost:3000/give-badge";
-  if (url) {
-    if (url.hostname === "localhost") {
-      prefixToGiveBadge = "http://localhost:3000/give-badge";
-    } else if (url.hostname === "trustful") {
-      prefixToGiveBadge = "https://trustful.vercel.app/give-badge";
-    }
-  }
-  const linkToGiveBadgeAddress = `${prefixToGiveBadge}?address=${address}`;
 
   return (
     <Flex flexDirection="column" minHeight="100vh" marginBottom="60px">
@@ -80,17 +68,7 @@ export const ShareSection = () => {
             >
               {address && chain ? (
                 <>
-                  <Canvas
-                    text={linkToGiveBadgeAddress}
-                    options={{
-                      errorCorrectionLevel: "M",
-                      width: 250,
-                      color: {
-                        dark: "#F5FFFFDC",
-                        light: "#212223",
-                      },
-                    }}
-                  />
+                  <QRCodeGiveBadge />
                   <Flex
                     color="white"
                     className="justify-center items-center gap-2"
